refactor: 출력 기능은 OutputView에서 전담할 수 있도록 리팩토링
