+++ src/main/java/lotto/view/OutputView.java
@@ -1,6 +1,38 @@
 package lotto.view;
 
+import lotto.domain.*;
+
+import java.util.List;
+
+import static lotto.domain.LottoConstant.LOTTO_SIZE;
+
 public class OutputView {
+    public static void printMyLotto(PurchasedLottos lottos) {
+        List<Lotto> lottoList = lottos.getLottoList();
+        for (Lotto lotto : lottoList) {
+            OutputView.printMessage(lotto.toString());
+        }
+        OutputView.printLine();
+    }
+
+    public static void showLottoStatistics(PurchasedLottos lottos, String lastWinningNumbers) {
+        List<Ranking> rankings = lottos.compareLottos(new Lotto(lastWinningNumbers));
+        LottoResult result = new LottoResult(rankings);
+        OutputView.printMessage("당첨 통계");
+        OutputView.printMessage("---------");
+        for (int matching = 3; matching <= LOTTO_SIZE; matching++) {
+            Ranking rank = Ranking.findRank(matching);
+            OutputView.printMessage("%d개 일치 (%d원)- %d개\r\n", matching, rank.getReward(), result.findRankings(matching).size());
+        }
+    }
+
+    public static void showLottoProfit(PurchasedLottos lottos, String lastWinningNumbers, Money money) {
+        List<Ranking> rankings = lottos.compareLottos(new Lotto(lastWinningNumbers));
+        LottoResult result = new LottoResult(rankings);
+        double profit = (double) result.calculateWinningMoney() / money.getMoney();
+        OutputView.printMessage("총 수익률은 %.2f입니다.", profit);
+    }
+
     public static void printMessage(String message) {
         System.out.println(message);
     }
